requirements.txt
azure-identity==1.25.1
azure-monitor-ingestion==1.1.0